Add files via upload
This can still be considered an early stage of development. The system for displaying images on the screen, such as the character, background, and text box, has been created, along with the text appearance system. This is done asynchronously, similar to how the "Dialogic" resource for VN creation in Godot works, but without the visual appearance achieved by that project’s plugins. (Additionally, I plan to take inspiration from it without copying to avoid any issues in the future.) This project consists purely of code.

Lastly, these asynchronous calls are executed from the "Scenery" node within its assigned script file. Before I forget, in this uploaded version, I am making progress on the save and load system. Over time, I will continue advancing to complete it and move on to the next stage.